requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python
numpy
scikit-learn
tqdm
matplotlib
pillow
